requirements.txt
appdirs==1.4.4
asgiref==3.4.1        
auth==0.5.3
blinker==1.4
certifi==2021.5.30    
chardet==4.0.0        
click==8.0.1
colorama==0.4.4       
decorator==5.0.9      
defusedxml==0.7.1     
distlib==0.3.2        
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.8.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
dnspython==2.1.0
dynaconf==3.1.7
eventlet==0.32.0
falcon==3.0.1
filelock==3.0.12
Flask==2.0.1
greenlet==1.1.1
gunicorn==20.1.0
hvac==0.11.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
knox==0.1.14
loguru==0.5.3
MarkupSafe==2.0.1
mongoengine==0.23.1
oauthlib==3.1.1
Pillow==8.3.1
psycopg2==2.8.6
pymongo==3.12.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
readchar==2.0.1
requests==2.25.1
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
startapp==0.2.0.0
urllib3==1.26.6
validators==0.18.2
virtualenv==20.4.7
Werkzeug==2.0.1
win32-setctime==1.0.3